:bug: Fix plugin-vibrant-image
The original plugin added the colors field to the node directly but
it was being flakey when run. According to the docs the node is immutable
so we shouldn't be manipulating it. Either I could use createNodeField
to add it under the fields key or make a separate node type and connect
it as a child of the File, which is what I chose. Now it needs to be
queried with childImageColors instead of just colors. This is the same
format as childImageSharp.

See:
- https://www.gatsbyjs.com/docs/node-creation/#changing-a-nodes-fields
- https://www.gatsbyjs.com/docs/reference/config-files/actions/#createNodeField
- https://www.gatsbyjs.com/docs/reference/config-files/actions/#createParentChildLink