:bug: Remove GatsbyImageColors fragment
This was erroring really inconsitently and I couldn't figure out why.
Gatsby wasn't picking up the fragment file from the plugin folder, I'm
not sure what the copy to cache was doing, and even adding it to the
src/fragments folder wasn't working. It could be because it's defined
more than once? I couldn't find any docs on how to include fragments in
plugins and nothing about the weird cache copy bit; decided it was just
easier to give up and query the fields manually.