:bug: Fix deployment authentication
I tried doing the build in the action but it was having a problem
with absolute paths. I tried adding the github token to the deploy
script and couldn't get it to work, so it's now a mishmash between
a deploy-prepare script and a general github action for deploying
to pages.

https://github.com/marketplace/actions/github-pages-action