:bug: Fix gatsby-plugin-vibrant-image
This was really messing up local dev. It complained that there wasn't
a childImageColors on File and I couldn't find it in graphiql, but
still it would render properly? I figured out it was because we're
trying to create the node inside the Vibrant promise instead of
awaiting the return of that data and doing it after. I also updated
some of the Gatsby APIs to match other plugins more closely.