:bug: Fix video ownedBy field
On BookCards the video review field stopped showing up. I suspected
it was because the related field is called 'ownedBy' which overrides
the 'ownedBy' text field in the markdown. GraphQL was getting
confused so I changed the related field to video.book and that's
fixed everything. Apart from mysteriously the video.book on the about
page.

This has also :zap: spead up the node transformations because to
relate a video to a book we're now using the much more efficient
findOne method. So so much faster.