:wrench: Improve diff for nightly deploys
Every night we redeploy with gh-pages but since it's saved in git, the diffs
are bloating the whole project. Thankfully on a redeploy where only the youtube
stats have changed, only the page-data files are edited, but since they're
minified it's quite a big diff. This prettifies the page data files during the
build process, so the nightly builds will cause minimal diffs.