:wrench: Force orphan commit for clean gh-pages deployments
The nightly deploys bloat the gh-pages branch with commits with huge
changes, because the youtube stats change a lot of numbers across the
site. I've tried cutting the diffs down by prettifying the files before
commit so the diffs are only for relevant info, but it's still in the
thousands.

This change should overwrite the branch data rather than adding to it.
I can't see any advantage of having a history of deploys.